HOLY CRAP WORKING SOMETHING
OK, here goes.  This does a very simple thing: it creates a socket to
the endpoint specified in the directive, sends a small message, recieves
a small message, and outputs it to the log.

Sample nginx configuration
        location /zmq/{
            zmq;
            zmq_endpoint "tcp://localhost:5555";
        }

a simple "hello world" server gets a response, and sends back the string
"World".  The resuliting output in the error log is:

Got a reply, it is: World